requirements.txt
environs==9.3.2
django==3.2
psycopg2==2.8.6
djangorestframework==3.12.4
django-filter==2.4.0
drf-yasg==1.20.0
djoser==2.1.0
djangorestframework-simplejwt==4.6.0
django-bootstrap4==3.0.1
asgiref==3.3.4
pytz==2021.1
gunicorn==20.1.0
